backend/requirements.txt
pygame
twilio
pyautogui
selenium
spotipy
SpeechRecognition
requests
pluggy
pytest
iniconfig
noisereduce
gtts
numpy
redis
urllib3
idna
certifi
joblib
tqdm
scipy
pylint
python-dotenv
pycaw